+++ libs/frontend/modules/builder/src/renderer/pipes/test/propMapBindingsPipe.spec.ts
@@ -1,7 +1,7 @@
 import { RenderProps } from '../../../store'
 import { propMapBindingsPipe } from '../propMapBindingsPipe'
 import { RenderContext } from '../types'
-import { elementToRender } from './data'
+import { elementToRender } from './renderData'
 import { EndPipeOutput } from './types'
 import { endPipe } from './utils'
