+++ libs/frontend/modules/builder/src/renderer/pipes/test/propTransformationJsPipe.spec.ts
@@ -1,7 +1,7 @@
 import { RenderProps } from '../../../store'
 import { propTransformationJsPipe } from '../propTransformationJsPipe'
 import { RenderContext } from '../types'
-import { elementToRender } from './data'
+import { elementToRender } from './renderData'
 import { EndPipeOutput } from './types'
 import { endPipe } from './utils'
