+++ libs/frontend/modules/builder/src/renderer/pipes/test/typedPropsPipe.spec.ts
@@ -1,6 +1,6 @@
 import { typedPropsPipe } from '../typedPropsPipe'
 import { RenderContext } from '../types'
-import { elementToRender } from './data'
+import { elementToRender } from './renderData'
 import { EndPipeOutput } from './types'
 import { endPipe } from './utils'
